refactor(codec): extract dispatch emitters into CodecEmitterDispatch.kt (step 4)
Move the sealed-dispatch emit cluster (buildDispatch*, buildFramedByDispatchOn*,
buildDispatchOnAggregating*, plus their DispatchShape/DispatchVariant/
PayloadCodecSource/LengthSource extension helpers), the shared
appendPeekFixedScalar value-class peek reconstructor, and the withVisibility
TypeSpec.Builder helpers to package-internal top-level. All are stateless emit
helpers (no batchCounter/codeGenerator/logger), so call sites in CodecEmitter
resolve them unqualified and unchanged.

CodecEmitter.kt 6328 -> 5058 lines; CodecEmitterDispatch.kt 1294 lines.
Byte-identical codegen verified by CodecSnapshotTest.
